refactor: move server.js to app folder
